core+site: round 93 — AccessConstraint, CohortLedger, warning fixes
Kernel additions:

policy::access:
- Constraint primitive with StaffGender / DeviceMode / allowed_languages /
  required_role_code dimensions. Vendor-neutral; the kernel evaluates,
  the operator configures.
- Context struct carrying request-scope attributes.
- evaluate(Constraint, Context) -> Result{Decision, reason} returning
  allow when every configured dimension is satisfied; deny with the
  first failed dimension named otherwise.
- to_string for all enums.

telemetry::CohortLedger:
- Append-only longitudinal observation chain. Used by GenericGuard,
  TriageMate post-triage outcomes, and any product needing per-patient
  or per-metric time-series.
- append / append_n / for_patient / for_metric / in_window / latest /
  metrics / patients / snapshot / clear.
- AggregateStats {count, mean, stddev, min, max} with stats_for_metric
  and stats_for_patient_metric.

Bug fixes:
- test_runtime.cpp: removed three useless casts (std::size_t == std::size_t
  comparison) flagged by -Wuseless-cast.
- test_telemetry.cpp: removed unused after_first variable.
- test_audit.cpp: changed loop counter int -> size_t to silence
  -Wsign-conversion on `i * 4`.

Tests: 1437 → 1511 (74 new); assertions: 35401 → 35800.
- test_access_constraint.cpp (11 cases)
- test_cohort_ledger.cpp (15 cases)

Linter additions in flight (also committed in this round):
- consent.hpp gained more methods.
- emergency_override.cpp gained active_for_patient and other helpers.
- test_consent_artefact.cpp grew from 21 -> 40 cases.
- test_emergency_override.cpp grew similarly.

+++ tests/test_runtime.cpp
@@ -1854,7 +1854,7 @@ TEST_CASE("ledger_length: matches ledger().length() after appends") {
             [](std::string s)->Result<std::string>{ return s; }));
         REQUIRE(inf.value().commit());
     }
-    CHECK(rt.ledger_length() == static_cast<std::size_t>(rt.ledger().length()));
+    CHECK(rt.ledger_length() == rt.ledger().length());
     CHECK(rt.ledger_length() >= 3);
 }
 
@@ -2176,8 +2176,7 @@ TEST_CASE("dispatched_inferences: matches metrics().count(\"inference.attempts\"
     }
     auto r = rt.dispatched_inferences();
     REQUIRE(r);
-    CHECK(r.value() ==
-          static_cast<std::size_t>(rt.metrics().count("inference.attempts")));
+    CHECK(r.value() == rt.metrics().count("inference.attempts"));
     CHECK(r.value() == 3);
 }
 
@@ -3326,8 +3325,7 @@ TEST_CASE("counter_total: matches metrics().counter_total() exactly") {
         [](std::string s) -> Result<std::string> { return s; }));
     REQUIRE(inf.value().commit());
 
-    CHECK(rt.counter_total()
-          == static_cast<std::size_t>(rt.metrics().counter_total()));
+    CHECK(rt.counter_total() == rt.metrics().counter_total());
 }
 
 // ============== Runtime::is_chain_well_formed =============================
